Implemented the Meals ViewModel
